ENH: Use exact type checks and make Py(Int|Float)... subclasses
Just to make the fix the type hierarchy a bit up. On the other
hand this breaks new subclasses a bit (something that I am not
really sure I want in any case, e.g. for UnitDType to be a subclass
of BasicNumericDType, it seems sensible, but for casting a bit
strange?).

But generally this is tricky, since inheriting casting slots is
probably counterproductive. So the question is in what regard
inheritance should even work. It could inherit a lot of other
things in principle.